Fixed screen reader issues - Fixed extra nav bar issues for most pages
